One 57117 fastlane add tests

+++ lib/fastlane/plugin/dynatrace/helper/dynatrace_helper.rb
@@ -1,6 +1,9 @@
 require 'fastlane_core/ui/ui'
 require 'digest'
-require 'open-uri'
+require 'net/http'
+require 'tempfile'
+require 'uri'
+
 
 module Fastlane
   UI = FastlaneCore::UI unless Fastlane.const_defined?("UI")
@@ -26,6 +29,7 @@ module Fastlane
         if not response.kind_of? Net::HTTPSuccess
           error_msg = "Couldn't update #{dtxDssClientBin} (invalid response: #{response.message} (#{response.code})) for URL: #{self.to_redacted_api_token_string(clientUri)})"
           self.check_fallback_or_raise(dtxDssClientPath, error_msg)
+          return dtxDssClientPath
         end
 
         # parse body
@@ -59,30 +63,31 @@ module Fastlane
            File.exists?(dtxDssClientPath) and 
            File.read(versionFilePath) == remoteClientUrl and
            File.size(dtxDssClientPath) > 0)
-          # extract and save client
           updatedClient = false
 
-          # prevents from creating a StringIO object instead of FileIO in URI.open if <10kB
-          OpenURI::Buffer.send :remove_const, 'StringMax' if OpenURI::Buffer.const_defined?('StringMax')
-          OpenURI::Buffer.const_set 'StringMax', 0
+          # extract and save client
+          zipped_tmp = self.save_to_tempfile(remoteClientUrl)
+          if File.size(zipped_tmp) <= 0
+            error_msg = "Downloaded symbolication client archive is empty (0 bytes)."
+            self.check_fallback_or_raise(dtxDssClientPath, error_msg)
+            return dtxDssClientPath
+          end
 
           begin
-            URI.open(remoteClientUrl) do |zipped|
-              UI.message "Unzipping fetched file with MD5 hash: #{Digest::MD5.new << IO.read(zipped)}"
-              Zip::InputStream.open(zipped) do |unzipped|
-                entry = unzipped.get_next_entry
-                if (entry.name == dtxDssClientBin)
-                  # remove old client
-                  UI.message "Found a different remote #{dtxDssClientBin} client. Removing local version and updating."
-                  File.delete(versionFilePath) if File.exist?(versionFilePath)
-                  File.delete(dtxDssClientPath) if File.exist?(dtxDssClientPath)
+            UI.message "Unzipping fetched file with MD5 hash: #{Digest::MD5.new << IO.read(zipped_tmp)}"
+            Zip::InputStream.open(zipped_tmp) do |unzipped|
+              entry = unzipped.get_next_entry
+              if (entry.name == dtxDssClientBin)
+                # remove old client
+                UI.message "Found a different remote #{dtxDssClientBin} client. Removing local version and updating."
+                File.delete(versionFilePath) if File.exist?(versionFilePath)
+                File.delete(dtxDssClientPath) if File.exist?(dtxDssClientPath)
 
-                  # write new client
-                  File.write(versionFilePath, remoteClientUrl)
-                  IO.copy_stream(entry.get_input_stream, dtxDssClientPath)
-                  FileUtils.chmod("+x", dtxDssClientPath)
-                  updatedClient = true
-                end
+                # write new client
+                File.write(versionFilePath, remoteClientUrl)
+                IO.copy_stream(entry.get_input_stream, dtxDssClientPath)
+                FileUtils.chmod("+x", dtxDssClientPath)
+                updatedClient = true
               end
             end
           rescue Zip::DecompressionError, 
@@ -91,9 +96,11 @@ module Fastlane
                  Zip::EntryNameError, 
                  Zip::EntrySizeError, 
                  Zip::GPFBit3Error, 
-                 Zip::InternalError 
+                 Zip::InternalError,
+                 Zip::CompressionMethodError 
             error_msg = "Could not update/extract #{dtxDssClientBin}, please try again."
             self.check_fallback_or_raise(dtxDssClientPath, error_msg)
+            return dtxDssClientPath
           end
 
           if updatedClient
@@ -101,6 +108,7 @@ module Fastlane
           else
             error_msg = "#{dtxDssClientBin} not found in served archive, please try again."
             self.check_fallback_or_raise(dtxDssClientPath, error_msg)
+            return dtxDssClientPath
           end
         end
         return dtxDssClientPath
@@ -117,8 +125,8 @@ module Fastlane
       private
       def self.check_fallback_or_raise(fallback_client, error)
         UI.important "If this error persists create an issue on our Github project (https://github.com/Dynatrace/fastlane-plugin-dynatrace/issues) or contact our support at https://www.dynatrace.com/support/contact-support/."
+        UI.important error
         if File.exists?(fallback_client) and File.size(fallback_client) > 0
-          UI.important error
           UI.important "Using cached client: #{fallback_client}"
         else
           UI.important "No cached fallback found."
@@ -126,13 +134,26 @@ module Fastlane
         end
       end
 
+      # assumes the token parameter is appended last (there is only one parameter anyway)
       def self.to_redacted_api_token_string(url)
         urlStr = url.to_s
         str = "Api-Token="
         idx = urlStr.index(str)
-        token_len = urlStr.length - idx + str.length + 1
+        token_len = urlStr.length - idx - str.length
         urlStr[idx + str.length..idx + str.length + token_len] = "-" * token_len
         return urlStr
+      end
+
+      def self.save_to_tempfile(url)
+        uri = URI.parse(url)
+        Net::HTTP.start(uri.host, uri.port) do |http|
+          resp = http.get(uri.path)
+          file = Tempfile.new('foo', Dir.tmpdir, 'wb+')
+          file.binmode
+          file.write(resp.body)
+          file.flush
+          file
+        end
       end
     end
   end
